requirements.txt
python-telegram-bot==21.3
fastapi==0.111.0
uvicorn[standard]==0.30.1
apscheduler==3.10.4
httpx==0.27.0
